refactor: priceModule 경로 업데이트
